+++ src/actions/tags.js
@@ -1,0 +1,35 @@
+/*
+ This file is a part of libertysoil.org website
+ Copyright (C) 2016  Loki Education (Social Enterprise)
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU Affero General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU Affero General Public License for more details.
+
+ You should have received a copy of the GNU Affero General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+export const SET_USER_TAGS = 'SET_USER_TAGS';
+export const SET_USER_RECENT_TAGS = 'SET_USER_RECENT_TAGS';
+
+export function setUserTags(tags) {
+  return {
+    type: SET_USER_TAGS,
+    hashtags: tags.hashtags,
+    schools: tags.schools,
+    geotags: tags.geotags
+  };
+}
+
+export function setUserRecentTags(recent_tags) {
+  return {
+    type: SET_USER_RECENT_TAGS,
+    recent_tags
+  };
+}
